Phase8

+++ app/controllers/reservations_controller.rb
@@ -1,19 +1,20 @@
 class ReservationsController < ApplicationController
-  include ReservationsHelper
   def index
-    @reservations = policy_scope(Resrvation).all
+    @reservations = if params[:status] == 'active'
+                      policy_scope(Resrvation.active)
+                    else
+                      policy_scope(Resrvation.archived)
+                    end
   end
 
   def create
-    if session[:reservation_id]
+    if current_user.resrvations.active.any?
       flash[:warning] = 'You have already booked'
-      redirect_to reservations_path
+      redirect_to root_path
     else
       authorize Resrvation
       @reservation = Resrvation.create(reserv_params)
       if @reservation.save
-        session[:reservation_id] = @reservation.id
-        room_count(@reservation.room_id)
         redirect_to root_path
         flash[:success] = 'Your reservation was created'
       end
@@ -23,18 +24,20 @@ class ReservationsController < ApplicationController
   def destroy
     authorize Resrvation
     @reservation = Resrvation.find(params[:id])
-    room_id = @reservation.room_id
-    if @reservation.destroy
-      session[:reservation_id] = nil
-      room_count(room_id)
+    if @reservation.status == 'active'
+      if @reservation.destroy
+        redirect_to root_path
+        flash[:danger] = 'You have deleted your reservation'
+      end
+    else
+      flash[:warning] = "You can't to delete this reservation"
       redirect_to root_path
-      flash[:danger] = 'You have canceled your reservation'
     end
   end
 
   private
 
   def reserv_params
-    params.permit(:user_id, :room_id)
+    params.permit(:user_id, :room_id, :status)
   end
 end
